feat: add shared CRT utilities
